Added the changes from the old LasaurApp
On our lasersaur we have the power supply modulated using the analog pins. I did
a fast attempt at porting these changes to this version, and it seems to be
working. Ran a few tests, and things seem okay, but not tested fully.

Note: This commit has been rebased by Matt. He added a comment to a modification in config.h