Copyied config.driveboard1403.h to config.h
comments in flashy.py seem to indicate that it will copy
each config.driveboardXXX.h to config.h and compile all different firmware variants.  The script itself, however, does not seem to do this.

Our laser cutter uses config.driveboard1403.h  (I think that the 1403 refers to the way our electronics are setup and what components we used when building it).